+++ src/server/js/check_npms.js
@@ -10,6 +10,7 @@ if( false ){
 
 checkNpmVersions({
     'lodash': '^4.17.0',
+    'simpl-schema': '^3.4.7'
 },
     'pwix:collection-timestampable'
 );